pyproject.toml
[project]
name = "anno-117-optimizer"
version = "0.1.0"
description = "Anno 117 Optimizer"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
]
